♻️ Extract built-in tools to lib/tools/built-in.ts
Move the 4 core built-in tools (compareOptions, webSearch, fetchPage,
deepResearch) and createSearchKnowledgeTool factory from route.ts to
a dedicated module.

This reduces route.ts by ~200 lines and makes the tools reusable
elsewhere (future: preflight validation, testing, documentation).
